+++ src/gamspy/_model.py
@@ -314,7 +314,7 @@ class Model:
         self.container._synch_with_gams()
 
     def __repr__(self) -> str:
-        return f"Model(name={self.name}, problem={self.problem}, equations={self.equations}, sense={self.sense}, objective={self._objective_variable}, matches={self._matches}, limited_variables={self._limited_variables}"
+        return f"Model(name='{self.name}', problem='{str(self.problem)}', equations={self.equations}, sense='{str(self.sense)}', objective={self._objective_variable}, matches={self._matches}, limited_variables={self._limited_variables}"
 
     def __str__(self) -> str:
         return (