+++ src/gamspy/_symbols/set.py
@@ -552,6 +552,9 @@ class Set(gt.Set, operable.Operable, Symbol, SetMixin):
         is_miro_input: bool = False,
         is_miro_output: bool = False,
     ):
+        if (is_miro_input or is_miro_output) and name is None:
+            raise ValidationError("Please specify a name for miro symbols.")
+
         self._is_miro_input = is_miro_input
         self._is_miro_output = is_miro_output
 
@@ -687,7 +690,7 @@ class Set(gt.Set, operable.Operable, Symbol, SetMixin):
         self._winner = "gams"
 
     def __repr__(self) -> str:
-        return f"Set(name={self.name}, domain={self.domain})"
+        return f"Set(name='{self.name}', domain={self.domain})"
 
     def _singleton_check(
         self,
@@ -830,7 +833,7 @@ class Set(gt.Set, operable.Operable, Symbol, SetMixin):
         if self.is_singleton:
             output = f"Singleton {output}"
 
-        output += self._get_domain_str()
+        output += self._get_domain_str(self.domain_forwarding)
 
         if self.description:
             output += f' "{self.description}"'