+++ src/gamspy/_symbols/variable.py
@@ -185,6 +185,9 @@ class Variable(gt.Variable, operable.Operable, Symbol):
         uels_on_axes: bool = False,
         is_miro_output: bool = False,
     ):
+        if is_miro_output and name is None:
+            raise ValidationError("Please specify a name for miro symbols.")
+
         # miro support
         self._is_miro_output = is_miro_output
 
@@ -307,7 +310,7 @@ class Variable(gt.Variable, operable.Operable, Symbol):
         return expression.Expression(self, "=e=", other)
 
     def __repr__(self) -> str:
-        return f"Variable(name={self.name}, domain={self.domain}, type={self.type})"
+        return f"Variable(name='{self.name}', domain={self.domain}, type={self.type})"
 
     @property
     def T(self) -> implicits.ImplicitVariable:
@@ -800,7 +803,7 @@ class Variable(gt.Variable, operable.Operable, Symbol):
         """
         representation = self.name
         if self.domain:
-            representation += self._get_domain_str()
+            representation += self._get_domain_str(self.domain_forwarding)
 
         return representation
 
@@ -823,12 +826,7 @@ class Variable(gt.Variable, operable.Operable, Symbol):
 
         """
         output = self.type + " "
-
-        statement_name = self.name
-        if self.domain:
-            statement_name += self._get_domain_str()
-
-        output += f"Variable {statement_name}"
+        output += f"Variable {self.gamsRepr()}"
 
         if self.description:
             output += ' "' + self.description + '"'